+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Removed
+
+- [**Breaking**] stdout/stderr are no longer replayed. Only if a script is
+  actually running will it now produce output to those streams.
 
 ## [0.6.1] - 2022-06-15